+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## tip
 
+* FEATURE: implements the getLogRowContext method. See [this PR](https://github.com/VictoriaMetrics/victorialogs-datasource/pull/247).
 * FEATURE: add support for build the linux/s390x. Extend backend build process to add more architectures. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/250).
 * FEATURE: add configuration screen for Custom query parameters. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/252).